+++ app/Http/Controllers/Admin/AdminJobExperienceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Admin;
 
 use App\Http\Controllers\Controller;
+use App\Models\Job;
 use Illuminate\Http\Request;
 use App\Models\JobExperience;
 
@@ -55,6 +56,10 @@ class AdminJobExperienceController extends Controller
 
     public function delete($id)
     {
+        $check = Job::where('job_experience_id',$id)->count();
+        if($check>0) {
+            return redirect()->back()->with('error', 'item is still active at other sections');
+        }
         JobExperience::where('id',$id)->delete();
         return redirect()->route('admin_job_experience_index')->with('success', 'Deleted Successfully.');
     }
